+++ src/settings/SettingsTabContext.ts
@@ -1,13 +1,17 @@
-import { App } from "obsidian";
+import type { App } from "obsidian";
 
-import { IVaultIntelligencePlugin } from "./types";
+import type { VaultIntelligenceSettingTab } from "./settingsTab";
+import type { IVaultIntelligencePlugin } from "./types";
 
 /**
  * Context object passed to settings tab render functions
- * Provides access to app, plugin, and the container element for the tab
+ * Provides access to app, plugin, the container element for the tab, and
+ * an optional reference to the parent setting tab instance for version-gated
+ * UI refresh (see refreshSettings/reRenderSection).
  */
 export interface SettingsTabContext {
     app: App;
     containerEl: HTMLElement;
     plugin: IVaultIntelligencePlugin;
+    tabInstance?: VaultIntelligenceSettingTab;
 }
